requirements.txt
httpx
uvicorn
fastapi
pydantic
python-dotenv
protobuf
aiosqlite
fake-useragent